+++ sdk/rpc-gql-compat/src/index.ts
@@ -1,6 +1,4 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { GraphQLSuiClient } from './client.js';
-
-export { GraphQLSuiClient };
+export * from './transport.js';
